84. Undefined behaviour in EntityAppearance class
Refactored the EntityAppearance class to eliminate the union and to
simply store the appearance code inside a KUINT32.  Transitions between
this basic datatype and the specific appearance structs are done via
reinterpret_cast.

This could probably be simplified further by simply changing the
specific appearance struct into classes and creating a ctor that takes a
KUINT32, but this should suffice to fix this bug for now.