Made more constructors explicit and expanded some enumerations.
Most PDU changes ware the addition of explicit to single parameter constructors. This avoids issues where its possible to implicitly convert most pdus into nearly any other pdu with a Header constructor.
Added OtherCamouflage to the EntityCamouflageDescriptor. Unclear when this was added so I cant quote a spec number.
Fixed what appears to be a spelling error in the KDefines that would probably prevent the variable its attempting to define working correctly.